Move test schemas to separate modules
