+++ tests/schemas/parent_to_child.py
@@ -1,0 +1,37 @@
+from __future__ import annotations
+
+from typing import (
+    Annotated,
+    TYPE_CHECKING,
+)
+
+from pydantic import ConfigDict
+from fastapi_jsonapi.schema_base import (
+    BaseModel,
+)
+from fastapi_jsonapi.types_metadata import RelationshipInfo
+
+if TYPE_CHECKING:
+    from tests.schemas.child import ChildSchema
+    from tests.schemas.parent import ParentSchema
+
+
+class ParentToChildAssociationAttributesSchema(BaseModel):
+    extra_data: str
+    model_config = ConfigDict(from_attributes=True)
+
+
+class ParentToChildAssociationSchema(ParentToChildAssociationAttributesSchema):
+    parent: Annotated[
+        ParentSchema | None,
+        RelationshipInfo(
+            resource_type="parent",
+        ),
+    ] = None
+
+    child: Annotated[
+        ChildSchema | None,
+        RelationshipInfo(
+            resource_type="child",
+        ),
+    ] = None
